refactor appstate
