+++ test/services/Levels.spec.js
@@ -1,73 +1,23 @@
-import LevelsService from '../../src/services/Levels.js'
+import ModesService from '../../src/services/Modes.js'
 import { React, TestUtils, fixtures, testdom } from '../react-helpers';
 require("babel/polyfill");
 
-describe.only("Levels Service", () => {
-  let Levels;
-  let appState;
-  let result;
+describe("Modes Service", () => {
+  let Modes;
 
   beforeEach(() => {
-    Levels = new LevelsService();
+    Modes = new ModesService();
   });
 
-  describe("set viewModel", () => {
-    let current;
-    let expectation;
-    let proposal;
+  it("sets the current mode", () => {
+    let expectation = {
+      current: 'learn',
+      learn: 'active',
+      curate: ''
+    };
 
-    describe("selection is child", () => {
-      beforeEach(() => {
+    let modes = Modes.set('learn');
 
-        current = {
-          waypoints: { id:1, title: 'home' },
-          waypoint: { id: 1, title: 'tada' },
-          checkpoint: { id: 1, title: 'tada' },
-          resource: false
-        };
-
-        expectation = {
-          current: 'resource',
-          waypoints: { id:1, title: 'home' },
-          waypoint: { id: 1, title: 'tada' },
-          checkpoint: { id: 1, title: 'tada' },
-          resource: { id: 1, title: 'yo' }
-        };
-
-        let selection = { type: 'resource', title: 'yo', id: 1 };
-        proposal = Levels.set({ current, selection });
-      });
-
-      it("sets the model to the proposal", () => {
-        expect(proposal).to.deep.equal(expectation);
-      });
-    });
-
-    describe("selection is parent", () => {
-      beforeEach(() => {
-
-        current = {
-          waypoints: { id:1, title: 'home' },
-          waypoint: { id: 1, title: 'tada' },
-          checkpoint: { id: 1, title: 'tada' },
-          resource: { id: 1, title: 'tada' }
-        };
-
-        expectation = {
-          current: 'waypoint',
-          waypoints: { id:1, title: 'home' },
-          waypoint: { id: 1, title: 'tada' },
-          checkpoint: false,
-          resource: false
-        };
-
-        let selection = { type: 'waypoint', title: 'tada', id: 1 };
-        proposal = Levels.set({ current, selection });
-      });
-
-      it("sets the model to the proposal", () => {
-        expect(proposal).to.deep.equal(expectation);
-      });
-    });
+    expect(modes).to.deep.equal(expectation);
   });
 });
